#16: rename modalphotopage.dart to photopage.dart
