+++ lib/Widgets/postmodal.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
-import 'package:re_frame/Widgets/modalphotopage.dart';
+import 'package:re_frame/Widgets/photopage.dart';
 import 'package:re_frame/calendar_util.dart';
 
 class PostModal extends StatelessWidget {
